+++ src/gate_editor/gates/gate_single/line_gate.rs
@@ -170,6 +170,12 @@ impl DrawableGate for LineGate {
         Box::new(self.clone())
     }
 
+    fn with_new_id(&self, new_id: Arc<str>) -> Option<Box<dyn DrawableGate>> {
+        let mut copy = self.clone();
+        copy.inner.id = new_id;
+        Some(Box::new(copy))
+    }
+
     fn get_id(&self) -> Arc<str> {
         self.inner.id.clone()
     }